Update EMGAnalysisV1.ino
I removed noTone because we already set the duration of the tone to be 250ms. However, I found online that it is required to have the delay since the Arduino would continue with the code if the delay was not there and there could be issues with the Arduino trying to play 2 tones at the same time.